requirements.txt
tqdm==4.23.0
pandas==0.22.0
imageio==2.2.0
pydot_ng==1.0.0
opencv_python==3.4.0.12
torch==0.3.1
attrdict==2.0.0
scipy==1.0.0
scikit_image==0.13.1
numpy==1.14.0
imgaug==0.2.5
neptune-cli
ipython==6.3.1
Pillow==5.1.0
PyYAML==3.12
Cython==0.28.2
pycocotools==2.0.0
steppy==0.1.14
steppy-toolkit==0.1.12
pretrainedmodels==0.7.0
torchvision==0.2.0

